Refactored bulk models into xml folder, removed unused references in cs files in the Common Libraries project.
