+++ src/CommonLibrariesForNET/AuthenticationClient.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Salesforce.Common.Models;
 using Newtonsoft.Json;
 using Salesforce.Common.Models.Json;
