+++ src/CommonLibrariesForNET/Common.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Salesforce.Common.Models;
 using Salesforce.Common.Models.Json;
 
 namespace Salesforce.Common
