+++ src/CommonLibrariesForNET/IServiceHttpClient.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using Salesforce.Common.Models;
 using Salesforce.Common.Models.Json;
 
 namespace Salesforce.Common
