+++ src/CommonLibrariesForNET/Models/Xml/BatchResult.cs
@@ -1,6 +1,6 @@
 ﻿using System.Xml.Serialization;
 
-namespace Salesforce.Force.Bulk.Models
+namespace Salesforce.Common.Models.Xml
 {
     [XmlType(TypeName = "result")]
     public class BatchResult
