+++ src/CommonLibrariesForNET/Models/Xml/BatchResultList.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Xml.Serialization;
 
-namespace Salesforce.Force.Bulk.Models
+namespace Salesforce.Common.Models.Xml
 {
     [XmlRoot(ElementName = "results",
         Namespace = "http://www.force.com/2009/06/asyncapi/dataload")]
